TeamSync/requirements.txt
redis-server
channels-redis 